i completed backend codes

+++ backend/src/main/java/com/backend/repositories/RepositoryAuthenticate.java
@@ -6,6 +6,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RepositoryAuthenticate extends JpaRepository<Users, Integer> {
+public interface RepositoryAuthenticate extends JpaRepository<Users, Long> {
 
 }
